deleted GeneriekeZaakDossierPage and moved it all to GzacBasePage
